Fill the authors section of all pages (#249)
People are only generally added if they were the main author of a PR (a suggested change doesn't count). People who made short additions have been added, if it still exists on the page. This brings an interesting question for this feature - if someone contributed to a page but their contribution isn't on there anymore then they probably shouldn't be on the authors list - but this is hard to keep track of!
Can close #207 when this is merged.

Co-authored-by: LachlanCourt <80299053+LachlanCourt@users.noreply.github.com>