Adds a NUClear tutorial (#262)
Adds a NUClear tutorial. Simple ping-pong program. 

I have ideas for expanding it to include utilities and testing, but not
100% sure on what to make the utility about. Basic idea was to have an
add function. Would also be good to expand at some point to using Eigen.

I'll have a review of this myself after today, I'm sure I've left in
references to previous version of this (changed the direction a couple
of times as I went).


https://deploy-preview-262--nubook.netlify.app/guides/main/nuclear-tutorial

Co-authored-by: Cameron Murtagh <43051455+CMurtagh-LGTM@users.noreply.github.com>
Co-authored-by: LachlanCourt <80299053+LachlanCourt@users.noreply.github.com>